split out affine opts
