+++ vyper/venom/optimization_levels/O2.py
@@ -10,6 +10,7 @@ from vyper.venom.passes import (
     CSE,
     SCCP,
     AlgebraicOptimizationPass,
+    AffineFoldingPass,
     AssertCombinerPass,
     AssignElimination,
     BranchOptimizationPass,
@@ -41,6 +42,7 @@ PASSES_O2: List[PassConfig] = [
     SimplifyCFGPass,
     MakeSSA,
     PhiEliminationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     SCCP,
     SimplifyCFGPass,
@@ -51,6 +53,7 @@ PASSES_O2: List[PassConfig] = [
     SCCP,
     SimplifyCFGPass,
     AssignElimination,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     LoadElimination,
     PhiEliminationPass,
@@ -80,6 +83,7 @@ PASSES_O2: List[PassConfig] = [
     LoadElimination,
     RemoveUnusedVariablesPass,
     BranchOptimizationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     AssertCombinerPass,
     # This improves the performance of cse
