+++ vyper/venom/optimization_levels/O3.py
@@ -10,6 +10,7 @@ from vyper.venom.passes import (
     CSE,
     SCCP,
     AlgebraicOptimizationPass,
+    AffineFoldingPass,
     AssertCombinerPass,
     AssertEliminationPass,
     AssignElimination,
@@ -45,6 +46,7 @@ PASSES_O3: List[PassConfig] = [
     SimplifyCFGPass,
     MakeSSA,
     PhiEliminationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     SCCP,
     SimplifyCFGPass,
@@ -55,6 +57,7 @@ PASSES_O3: List[PassConfig] = [
     SCCP,
     SimplifyCFGPass,
     AssignElimination,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     AssertEliminationPass,
     OverflowEliminationPass,
@@ -87,6 +90,7 @@ PASSES_O3: List[PassConfig] = [
     MemoryCopyElisionPass,
     RemoveUnusedVariablesPass,
     BranchOptimizationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     AssertCombinerPass,
     # This improves the performance of cse
