+++ vyper/venom/optimization_levels/Os.py
@@ -10,6 +10,7 @@ from vyper.venom.passes import (
     CSE,
     SCCP,
     AlgebraicOptimizationPass,
+    AffineFoldingPass,
     AssertCombinerPass,
     AssignElimination,
     BranchOptimizationPass,
@@ -42,6 +43,7 @@ PASSES_Os: List[PassConfig] = [
     SimplifyCFGPass,
     MakeSSA,
     PhiEliminationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     SCCP,
     SimplifyCFGPass,
@@ -52,6 +54,7 @@ PASSES_Os: List[PassConfig] = [
     SCCP,
     SimplifyCFGPass,
     AssignElimination,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     LoadElimination,
     PhiEliminationPass,
@@ -81,6 +84,7 @@ PASSES_Os: List[PassConfig] = [
     LoadElimination,
     RemoveUnusedVariablesPass,
     BranchOptimizationPass,
+    AffineFoldingPass,
     AlgebraicOptimizationPass,
     AssertCombinerPass,
     # This improves the performance of cse
