+++ vyper/venom/passes/__init__.py
@@ -1,4 +1,5 @@
 from .algebraic_optimization import AlgebraicOptimizationPass
+from .affine_folding import AffineFoldingPass
 from .assert_combiner import AssertCombinerPass
 from .assert_elimination import AssertEliminationPass
 from .assign_elimination import AssignElimination
